requirements.txt
# Core
ccxt==4.4.36
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1

# Technical Analysis
ta==0.11.0

# Machine Learning
scikit-learn==1.5.2
joblib==1.4.2
xgboost==2.1.3

# Dashboard
streamlit==1.41.1
plotly==5.24.1

# Notifications
python-telegram-bot==21.9
aiohttp==3.11.11

# Market Data (Stock/ETF)
yfinance>=0.2.36
requests>=2.31.0
beautifulsoup4>=4.12.0
scipy>=1.11.0

# Utilities
schedule==1.2.2
loguru==0.7.3
rich==13.9.4
